+++ packages/easyops-analytics/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./ApiAnalysis";
+export * from "./userAnalytics";