+++ packages/editor-bricks-helper/src/internal/BuilderDataManager.spec.ts
@@ -138,7 +138,6 @@ Object {
         "brick-a",
       ],
       "$$normalized": Object {
-        "alias": "alias-a",
         "brick": "brick-a",
       },
       "$$parsedEvents": Object {},
@@ -868,7 +867,6 @@ Object {
         "brick-a",
       ],
       "$$normalized": Object {
-        "alias": "alias-a",
         "brick": "brick-a",
       },
       "$$parsedEvents": Object {},