+++ dll/react-dnd/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.181](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.180...@next-dll/react-dnd@0.1.181) (2021-10-22)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+
+
+
+
+## [0.1.180](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.179...@next-dll/react-dnd@0.1.180) (2021-10-21)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+
+
+
+
 ## [0.1.179](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.178...@next-dll/react-dnd@0.1.179) (2021-10-20)
 
 **Note:** Version bump only for package @next-dll/react-dnd