+++ packages/brick-kit/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.78.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.78.0...@next-core/brick-kit@2.78.1) (2021-10-22)
+
+**Note:** Version bump only for package @next-core/brick-kit
+
+
+
+
+
+# [2.78.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.77.0...@next-core/brick-kit@2.78.0) (2021-10-21)
+
+
+### Features
+
+* **bindListeners:** add analytics.send builtin handler ([bd16412](https://github.com/easyops-cn/next-core/commit/bd164129a1e95b73867be9504c0704beb14fe581))
+
+
+
+
+
 # [2.77.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.76.0...@next-core/brick-kit@2.77.0) (2021-10-20)